dns_server: Put more code in common
This will allow a python module to be written to modify DNS entries in sam.ldb directly
